requirements.txt
fastapi
uvicorn[standard]
requests
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dateutil
openai